Move pauseimage into its own package
This avoids a circular import in the dockertools tests
