+++ pkg/kubelet/network/cni/cni_test.go
@@ -28,12 +28,12 @@ import (
 	"testing"
 	"text/template"
 
+	"k8s.io/kubernetes/cmd/kubelet/app/options/pauseimage"
 	clientset "k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset"
 
 	docker "github.com/fsouza/go-dockerclient"
 	cadvisorapi "github.com/google/cadvisor/info/v1"
 
-	"k8s.io/kubernetes/cmd/kubelet/app/options"
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/client/record"
 	kubecontainer "k8s.io/kubernetes/pkg/kubelet/container"
@@ -152,7 +152,7 @@ func newTestDockerManager() (*dockertools.DockerManager, *dockertools.FakeDocker
 		proberesults.NewManager(),
 		containerRefManager,
 		&cadvisorapi.MachineInfo{},
-		options.GetDefaultPodInfraContainerImage(),
+		pauseimage.GetDefaultPodInfraContainerImage(),
 		0, 0, "",
 		containertest.FakeOS{},
 		networkPlugin,
